+++ Source/SIMPLib/Geometry/TriangleGeom.h
@@ -143,14 +143,14 @@ public:
    * @param triId
    * @param verts
    */
-  void setVertsAtTri(size_t triId, int64_t verts[3]);
+  void setVertsAtTri(size_t triId, size_t verts[3]);
 
   /**
    * @brief getVertsAtTri
    * @param triId
    * @param verts
    */
-  void getVertsAtTri(size_t triId, int64_t verts[3]);
+  void getVertsAtTri(size_t triId, size_t verts[3]);
 
   /**
    * @brief getVertCoordsAtTri
@@ -166,7 +166,7 @@ public:
    * @param i
    * @return
    */
-  int64_t* getTriPointer(size_t i);
+  size_t* getTriPointer(size_t i);
 
   /**
    * @brief getNumberOfTris
@@ -386,14 +386,14 @@ public:
    * @param edgeId
    * @param verts
    */
-  void setVertsAtEdge(size_t edgeId, int64_t verts[2]) override;
+  void setVertsAtEdge(size_t edgeId, size_t verts[2]) override;
 
   /**
    * @brief getVerts
    * @param edgeId
    * @param verts
    */
-  void getVertsAtEdge(size_t edgeId, int64_t verts[2]) override;
+  void getVertsAtEdge(size_t edgeId, size_t verts[2]) override;
 
   /**
    * @brief getVertCoordsAtEdge
@@ -408,7 +408,7 @@ public:
    * @param i
    * @return
    */
-  int64_t* getEdgePointer(size_t i) override;
+  size_t* getEdgePointer(size_t i) override;
 
   /**
    * @brief getNumberOfEdges
